requirements.txt
urllib3
python-dotenv
pymisp
watchdog
tkinter
